fix(consensus): use u128/big.Int for subsidy arithmetic (§19.1) (#420)
* fix(consensus): use u128/big.Int for subsidy arithmetic (§19.1) (Q-AUDIT-POST-04)

* fix(rust): bridge u128 subsidy types at node/CLI boundary

Add u64↔u128 conversions at the boundary between consensus crate
(u128 internally) and node/CLI (u64 JSON/disk format). Fixes CI
build failure where rubin-node and rubin-consensus-cli were not
updated for the u128 already_generated type change.


---------

+++ clients/rust/crates/rubin-consensus/src/connect_block_inmem.rs
@@ -13,14 +13,14 @@ use crate::utxo_basic::{apply_non_coinbase_tx_basic_update_with_mtp, Outpoint, U
 pub struct InMemoryChainState {
     pub utxos: HashMap<Outpoint, UtxoEntry>,
     /// already_generated(h): subsidy-only (excluding fees).
-    pub already_generated: u64,
+    pub already_generated: u128,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct ConnectBlockBasicSummary {
     pub sum_fees: u64,
-    pub already_generated: u64,
-    pub already_generated_n1: u64,
+    pub already_generated: u128,
+    pub already_generated_n1: u128,
     pub utxo_count: u64,
 }
 
@@ -101,7 +101,7 @@ pub fn connect_block_basic_in_memory_at_height(
     if block_height != 0 {
         let subsidy = block_subsidy(block_height, already_generated);
         already_generated_n1 = already_generated
-            .checked_add(subsidy)
+            .checked_add(u128::from(subsidy))
             .ok_or_else(|| TxError::new(ErrorCode::BlockErrParse, "already_generated overflow"))?;
         state.already_generated = already_generated_n1;
     }
